% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPackageround4-package.R
\docType{package}
\name{myPackageround4-package}
\alias{myPackageround4}
\alias{myPackageround4-package}
\title{myPackageround4: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whitneysu/myPackageround4}
  \item Report bugs at \url{https://github.com/whitneysu/myPackageround4/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
